src/requirements.txt

psycopg2==2.6.1
Flask==0.11.1
facebook-sdk==1.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
requests==2.10.0